# api/app/db/generated/models.py
# Code generated by sqlc. DO NOT EDIT.
# versions:
#   sqlc v1.29.0
import datetime
import enum

import pydantic


class AvailabilityStatus(enum.StrEnum):
    ONLINE = "online"
    IDLE = "idle"
    DND = "dnd"
    OFFLINE = "offline"


class ChannelType(enum.StrEnum):
    TEXT = "text"
    VOICE = "voice"


class Channel(pydantic.BaseModel):
    id: str
    created: datetime.datetime
    modified: datetime.datetime
    name: str
    guild: str
    channel_group: str | None
    type: ChannelType
    is_general: bool
    order: int | None


class ChannelGroup(pydantic.BaseModel):
    id: str
    created: datetime.datetime
    modified: datetime.datetime
    name: str
    guild: str
    order: int


class Emoji(pydantic.BaseModel):
    id: str
    created: datetime.datetime
    modified: datetime.datetime
    name: str
    source: str
    guild: str
    author: str | None


class Guild(pydantic.BaseModel):
    id: str
    created: datetime.datetime
    modified: datetime.datetime
    name: str
    avatar: str | None
    banner: str | None


class Message(pydantic.BaseModel):
    id: str
    created: datetime.datetime
    modified: datetime.datetime
    author: str | None
    channel: str
    content: str
    is_edited: bool


class User(pydantic.BaseModel):
    id: str
    created: datetime.datetime
    modified: datetime.datetime
    username: str
    discriminator: int
    full_username: str
    email: str
    avatar: str | None
    cover: str | None
    availability_status: AvailabilityStatus


class UserGuildsOrder(pydantic.BaseModel):
    user: str
    guild: str
    order: int