# api_go/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Fastcord",
    "contact": {
      "name": "KVD Studio",
      "url": "https://kvd.studio",
      "email": "hello@kvd.studio"
    },
    "version": "0.1.0"
  },
  "host": "localhost:8000",
  "basePath": "/api",
  "paths": {
    "/guilds": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["guild"],
        "summary": "List all guilds",
        "parameters": [
          {
            "type": "integer",
            "description": "Page number",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Page size",
            "name": "page_size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/validation.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["guild"],
        "summary": "Create guild",
        "parameters": [
          {
            "description": "Guild details",
            "name": "guild",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.CreateGuildInput"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/sql.Guild"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/validation.ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "models.CreateGuildInput": {
      "type": "object",
      "required": ["name"],
      "properties": {
        "avatar": {
          "type": "string"
        },
        "banner": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "pgtype.InfinityModifier": {
      "type": "integer",
      "enum": [1, 0, -1],
      "x-enum-varnames": ["Infinity", "Finite", "NegativeInfinity"]
    },
    "pgtype.Text": {
      "type": "object",
      "properties": {
        "string": {
          "type": "string"
        },
        "valid": {
          "type": "boolean"
        }
      }
    },
    "pgtype.Timestamp": {
      "type": "object",
      "properties": {
        "infinityModifier": {
          "$ref": "#/definitions/pgtype.InfinityModifier"
        },
        "time": {
          "description": "Time zone will be ignored when encoding to PostgreSQL.",
          "type": "string"
        },
        "valid": {
          "type": "boolean"
        }
      }
    },
    "sql.Guild": {
      "type": "object",
      "properties": {
        "avatar": {
          "$ref": "#/definitions/pgtype.Text"
        },
        "banner": {
          "$ref": "#/definitions/pgtype.Text"
        },
        "created": {
          "$ref": "#/definitions/pgtype.Timestamp"
        },
        "id": {
          "type": "string"
        },
        "modified": {
          "$ref": "#/definitions/pgtype.Timestamp"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "validation.ErrorResponse": {
      "type": "object",
      "properties": {
        "actual": {},
        "expected": {
          "type": "string"
        },
        "field": {
          "type": "string"
        }
      }
    }
  }
}